Update systems.js
Added Crucible to the list of RPG systems, however not sure what the return is, I have tried to look in the console but I can't find it, it is called 'stride' in the game. They measure in squares and not feet or metres. It also needs to be tied to action points. SO much is free movement, then an AP per movement of the actor. But no problem if it can't be done yet.