Add terrain mapper calculation when calculating distance
Terrain Mapper assigns active effects to tokens as they move across the canvas. This addition uses the Terrain Mapper API to determine how the token movement is penalized or given a bonus based on terrain along the segment(s). That penalty or bonus is used to shrink or increase the perceived distance (from the token's frame of reference). 

For example, if the token  movement is penalized by 50%, distance is accordingly increased by 200% to account for this.