Fix check to enable/disable Terrain Ruler coloring/distance
`window.terrainRuler` is truthy if it exists, but it should check `window.terrainRuler.active` to make sure we actually want it on anyway.

This does have the minor effect of a "desync", where the colors/distance traveled are client-side based on whether the client user has terrain ruler enabled, so they might see something different from the dragger, but that's minimal.